+++ src/theme/decorations.js
@@ -1,6 +1,7 @@
 const decorations = {
   shadow: '0 2px 4px 0 rgba(49, 50, 51, 0.2)',
   borderRadius: '0.25rem',
+  borderRadiusLarge: '0.5rem',
   noShadow: 'none',
 };